Add tasks index to storage.
The tasks index will make it easier have access to the tasks,
instead previously we have to go through all of them to find the board_id.
We also cache the last id for both boards and tasks, to avoid the need to
go through all of them to find the it.
There were some what a major refactoring to adapt with the new storage
structure.
The version was also updated from 0.1.1 to 0.1.2.